Extract functionality for determining Xiaomi air filter type into a separate util class
