+++ miio/airpurifier_miot.py
@@ -1,10 +1,10 @@
 import enum
 import logging
-import re
 from typing import Any, Dict, Optional
 
 import click
 
+from .airfilter import FilterType, FilterTypeUtil
 from .click_common import EnumType, command, format_output
 from .device import DeviceException
 from .miot_device import MiotDevice
@@ -29,26 +29,11 @@ class LedBrightness(enum.Enum):
     Off = 2
 
 
-class FilterType(enum.Enum):
-    Regular = "regular"
-    AntiBacterial = "anti-bacterial"
-    AntiFormaldehyde = "anti-formaldehyde"
-    Unknown = "unknown"
-
-
-FILTER_TYPE_RE = (
-    (re.compile(r"^\d+:\d+:41:30$"), FilterType.AntiBacterial),
-    (re.compile(r"^\d+:\d+:(30|0|00):31$"), FilterType.AntiFormaldehyde),
-    (re.compile(r".*"), FilterType.Regular),
-)
-
-
 class AirPurifierMiotStatus:
     """Container for status reports from the air purifier."""
 
-    _filter_type_cache = {}
-
     def __init__(self, data: Dict[str, Any]) -> None:
+        self.filter_type_util = FilterTypeUtil()
         self.data = data
 
     @property
@@ -173,22 +158,12 @@ class AirPurifierMiotStatus:
             return FilterType.Unknown
         if self.filter_rfid_product_id is None:
             return FilterType.Regular
-        return self._get_filter_type(self.filter_rfid_product_id)
+        return self.filter_type_util.determine_filter_type(self.filter_rfid_product_id)
 
     @property
     def button_pressed(self) -> Optional[str]:
         """Last pressed button."""
         return self.data["button_pressed"]
-
-    @classmethod
-    def _get_filter_type(cls, product_id: str) -> FilterType:
-        ft = cls._filter_type_cache.get(product_id, None)
-        if ft is None:
-            for filter_re, filter_type in FILTER_TYPE_RE:
-                if filter_re.match(product_id):
-                    ft = cls._filter_type_cache[product_id] = filter_type
-                    break
-        return ft
 
     def __repr__(self) -> str:
         s = (
